+++ Changelog.md
@@ -11,6 +11,9 @@
 1. Depend on ign-transport8 and ign-msgs5
     * [Pull request 101](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/101)
 
+1. Add Thermal Camera Sensor
+    * [Pull request 100](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/100)
+
 1. Updating exports and includes
     * [Pull request 98](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/98)